Fix management of child requests in subjects.
There was an issue with the incorrect subscriptions being returned
from the subscription factory. The subscriptions for new requests
has been fixed and the main subscription is now managed and canceled
from unsubscribers.
Because of the way the subjects work in RxJava, a lot of work had
to be done to monitor the state of subscriptions inside of a
subject inside the `CompositeRequestManager` class.
Subjects were also changed to replaysubjects, since that makes more
sense for data lookups.